fixed for file name `[aaaa] bbbbb.ccc`
aria2 treat `[ ]` as loop download request. if any file url that contains `[ ]`, but doesn't match `[00-11]` or `[001-222]` or ... so on, aria2 rpc will return an error of `loop range error.` and the download won't be processed
this should have been fixed on aria2 side though. my method may be a crude work around